chore: moving publisher to events pkg
