+++ x/contracts/events/publisher/http_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/google/go-cmp/cmp"
-	"github.com/primevprotocol/mev-commit/x/contracts/publisher"
+	"github.com/primevprotocol/mev-commit/x/contracts/events/publisher"
 	"github.com/primevprotocol/mev-commit/x/util"
 )
